+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.0.0]
+
+### Added
+
+- Minimal Resize WebGL template
+- WebGL Importer
+
+### Changed
+
+- README.md
+
 ## [0.0.1] - 2024-12-14
 
 ### Added
@@ -9,4 +20,5 @@
 - License (GPL-3.0)
 
 [Unreleased]: https://github.com/lajawi/unity-webgl-templates/compare/v0.0.1...HEAD
+[1.0.0]: https://github.com/lajawi/unity-webgl-templates/compare/v0.0.1...v1.0.0
 [0.0.1]: https://github.com/lajawi/unity-webgl-templates/releases/tag/v0.0.1